More stuff in response to static inspection

+++ core/src/main/java/com/google/zxing/common/detector/MathUtils.java
@@ -45,8 +45,8 @@ public final class MathUtils {
    * @return Euclidean distance between points A and B
    */
   public static float distance(float aX, float aY, float bX, float bY) {
-    float xDiff = aX - bX;
-    float yDiff = aY - bY;
+    double xDiff = aX - bX;
+    double yDiff = aY - bY;
     return (float) Math.sqrt(xDiff * xDiff + yDiff * yDiff);
   }
 
@@ -58,8 +58,8 @@ public final class MathUtils {
    * @return Euclidean distance between points A and B
    */
   public static float distance(int aX, int aY, int bX, int bY) {
-    int xDiff = aX - bX;
-    int yDiff = aY - bY;
+    double xDiff = aX - bX;
+    double yDiff = aY - bY;
     return (float) Math.sqrt(xDiff * xDiff + yDiff * yDiff);
   }
